Move build ID functionality into separate module
Let's move the build ID functionality into a separate module, to improve
code organization.
